+++ Anim-Paint/src/Controls/Slider.cpp
@@ -61,7 +61,7 @@ void Slider::setPosition(sf::Vector2i position) {
 	_barRect.position = position;
 	_thumbRect.position = position;
 	
-	int text_margin = basicFont.getLineSpacing(slider_font_size) / 2;
+	int text_margin = (int)(basicFont.getLineSpacing)(slider_font_size) / 2;
 
 	sf::Vector2f nameTextPos;
 	nameTextPos.x = (float)(position.x + text_margin);
